Implement restaurant ownership and authorization features

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
@@ -10,14 +12,17 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
     {
         private readonly ILogger _logger;
         private readonly IRestaurantRepository _restaurantRepository;
+        private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
         public DeleteRestaurantCommandHandler(
             ILogger<DeleteRestaurantCommand> logger,
-            IRestaurantRepository restaurantRepository
+            IRestaurantRepository restaurantRepository,
+            IRestaurantAuthorizationService restaurantAuthorizationService
         )
         {
             _logger = logger;
             _restaurantRepository = restaurantRepository;
+            _restaurantAuthorizationService = restaurantAuthorizationService;
         }
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -27,6 +32,12 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             {
                 _logger.LogWarning("Restaurant with ID: {RestaurantId} not found", request.Id);
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+            }
+
+            if(! _restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            {
+                _logger.LogWarning("Unauthorized attempt to delete restaurant with ID: {RestaurantId}", request.Id);
+                throw new ForbidException();
             }
             
             await _restaurantRepository.DeleteAsync(restaurant);
